requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-filter==2.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
Markdown==3.1.1
pkg-resources==0.0.0
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
static3==0.7.0
